requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
Flask-Limiter==4.1.1
python-dotenv==1.0.1
qrcode[pil]==7.4.2
CairoSVG>=2.7.1,<3
google-api-python-client>=2.130.0,<3
google-auth>=2.29.0,<3
google-auth-oauthlib>=1.2.0,<2
psycopg[binary]>=3.2.0,<4
gunicorn>=22.0.0,<23
